optimize:move NewsItem & BottomNavigation  to components package
