+++ app/src/main/java/ir/ahe/abbas/newstest/main/components/NewsItem.kt
@@ -1,4 +1,4 @@
-package ir.ahe.abbas.newstest.main
+package ir.ahe.abbas.newstest.main.components
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
